Create ECG_AD8232.ino
This is the source code for Esp32 C3 mini oled wich need to be compiled in Arduino it was tested on IDE v. 2.3.6
It need following libraries U8g2lib, NimBLE-Arduino 2.3.4.
when you compile you need to specify ESP32C3 Dev Module and the afferent com port in order to flash it. 

Caution some modules need to press boot button and connect it to usb befor programming in order to enter in boot mode.
also  i use cristal of 40MHz and 160MHz Wifi .

if you cannot program check the com port and use the tool \AppData\Local\Arduino15\packages\esp32\tools\esptool_py\5.0.0/esptool.exe    in order to full erase the flash of the esp.